+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where translations weren't working
+
 ## 1.0.0 - 2022-09-12
 
 ### Added